Bijwerken

+++ examplesDotNET/PacmanTwo/AIController.java
@@ -8,7 +8,7 @@
  * [http://www.fsf.org/copyleft/lgpl.html]
  * 
  * @author Michiel Hendriks
- * @version $Id: AIController.java,v 1.5 2006/09/05 06:58:49 elmuerte Exp $
+ * @version $Id: AIController.java,v 1.1 2006/09/05 07:12:14 elmuerte Exp $
  */
 package PacmanTwo;
 
@@ -20,23 +20,23 @@ import PacmanTwo.Strategy.*;
 public class AIController extends Controller implements Tickable
 {
 	protected int direction;
-	protected double dircnt = 0;
+	protected float dircnt = 0;
 	protected java.util.Random random;
 
 	public AIController()
 	{
 		random = new java.util.Random();
-		dircnt = 3+random.nextDouble()*3;
+		dircnt = 3+random.nextFloat()*3;
 		Game.instance().addTickElement(this);
 	}
 
-	public void tick(double delta)
+	public void tick(float delta)
 	{
 		dircnt -= delta;
 		if (dircnt <= 0)
 		{
 			getNextMove();
-			dircnt = 1+random.nextDouble()*4;
+			dircnt = 1+random.nextFloat()*4;
 		}
 	}
